requirements.txt
streamlit
google-genai
dotenv
langchain_google_genai
langchain_chroma
langchain_community
scikit-learn